Replace #[cfg] with wasix-specific submodule
