+++ src/wasi/wasi.rs
@@ -1,0 +1,1 @@
+pub type sigset_t = c_uchar;
